requirements.txt
coverage==7.13.4
iniconfig==2.3.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
